+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/panels/UploadsButtonController.java
@@ -23,6 +23,7 @@ import org.slf4j.LoggerFactory;
 import java.lang.invoke.MethodHandles;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -62,6 +63,9 @@ public class UploadsButtonController implements Initializable {
 
   @FXML
   private MenuItem altColorUploadItem;
+
+  @FXML
+  private MenuItem fplItem;
 
   @FXML
   private MenuItem dmdUploadItem;
@@ -196,6 +200,11 @@ public class UploadsButtonController implements Initializable {
   }
 
   @FXML
+  public void onFplUpload() {
+    TableDialogs.directUpload(stage, AssetType.FPL, null, null);
+  }
+
+  @FXML
   public void onBamCfgUpload() {
     List<GameRepresentation> selectedItems = getSelections();
     if (selectedItems != null && !selectedItems.isEmpty()) {
@@ -270,6 +279,7 @@ public class UploadsButtonController implements Initializable {
     altSoundUploadItem.setVisible(vpxEmulator);
     altColorUploadItem.setVisible(vpxEmulator);
     dmdUploadItem.setVisible(vpxEmulator);
+    fplItem.setVisible(fpEmulator);
     patchItem.setVisible(vpxEmulator);
     iniUploadMenuItem.setVisible(vpxEmulator);
     povItem.setVisible(vpxEmulator);